requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2~=3.0.1
chromadb
faiss-cpu
langchain_google_genai
langchain-community